requirements.txt
asgiref==3.5.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
Django==3.2.12
ffmpeg==1.4
googletrans==3.1.0a0
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.2
lxml==4.8.0
packaging==21.3
Pillow==9.0.1
playsound==1.3.0
pydub==0.25.1
pyenchant==3.2.2
pyobjc==8.3
pyobjc-core==8.3
pyobjc-framework-Accounts==8.3
pyobjc-framework-AddressBook==8.3
pyobjc-framework-AppleScriptKit==8.3
pyobjc-framework-AppleScriptObjC==8.3
pyobjc-framework-ApplicationServices==8.3
pyobjc-framework-AudioVideoBridging==8.3
pyobjc-framework-Automator==8.3
pyobjc-framework-AVFoundation==8.3
pyobjc-framework-AVKit==8.3
pyobjc-framework-CalendarStore==8.3
pyobjc-framework-CFNetwork==8.3
pyobjc-framework-CloudKit==8.3
pyobjc-framework-Cocoa==8.3
pyobjc-framework-Collaboration==8.3
pyobjc-framework-ColorSync==8.3
pyobjc-framework-Contacts==8.3
pyobjc-framework-ContactsUI==8.3
pyobjc-framework-CoreAudio==8.3
pyobjc-framework-CoreAudioKit==8.3
pyobjc-framework-CoreBluetooth==8.3
pyobjc-framework-CoreData==8.3
pyobjc-framework-CoreLocation==8.3
pyobjc-framework-CoreMedia==8.3
pyobjc-framework-CoreMediaIO==8.3
pyobjc-framework-CoreMIDI==8.3
pyobjc-framework-CoreML==8.3
pyobjc-framework-CoreServices==8.3
pyobjc-framework-CoreSpotlight==8.3
pyobjc-framework-CoreText==8.3
pyobjc-framework-CoreWLAN==8.3
pyobjc-framework-CryptoTokenKit==8.3
pyobjc-framework-DictionaryServices==8.3
pyobjc-framework-DiscRecording==8.3
pyobjc-framework-DiscRecordingUI==8.3
pyobjc-framework-DiskArbitration==8.3
pyobjc-framework-DVDPlayback==8.3
pyobjc-framework-EventKit==8.3
pyobjc-framework-ExceptionHandling==8.3
pyobjc-framework-ExternalAccessory==8.3
pyobjc-framework-FinderSync==8.3
pyobjc-framework-FSEvents==8.3
pyobjc-framework-GameCenter==8.3
pyobjc-framework-GameController==8.3
pyobjc-framework-GameKit==8.3
pyobjc-framework-GameplayKit==8.3
pyobjc-framework-ImageCaptureCore==8.3
pyobjc-framework-IMServicePlugIn==8.3
pyobjc-framework-InputMethodKit==8.3
pyobjc-framework-InstallerPlugins==8.3
pyobjc-framework-InstantMessage==8.3
pyobjc-framework-Intents==8.3
pyobjc-framework-IOSurface==8.3
pyobjc-framework-iTunesLibrary==8.3
pyobjc-framework-LatentSemanticMapping==8.3
pyobjc-framework-LaunchServices==8.3
pyobjc-framework-libdispatch==8.3
pyobjc-framework-LocalAuthentication==8.3
pyobjc-framework-MapKit==8.3
pyobjc-framework-MediaAccessibility==8.3
pyobjc-framework-MediaLibrary==8.3
pyobjc-framework-MediaPlayer==8.3
pyobjc-framework-MediaToolbox==8.3
pyobjc-framework-Metal==8.3
pyobjc-framework-MetalKit==8.3
pyobjc-framework-MetalPerformanceShaders==8.3
pyobjc-framework-ModelIO==8.3
pyobjc-framework-MultipeerConnectivity==8.3
pyobjc-framework-NetFS==8.3
pyobjc-framework-NetworkExtension==8.3
pyobjc-framework-NotificationCenter==8.3
pyobjc-framework-OpenDirectory==8.3
pyobjc-framework-OSAKit==8.3
pyobjc-framework-Photos==8.3
pyobjc-framework-PhotosUI==8.3
pyobjc-framework-PreferencePanes==8.3
pyobjc-framework-PubSub==8.3
pyobjc-framework-Quartz==8.3
pyobjc-framework-SafariServices==8.3
pyobjc-framework-SceneKit==8.3
pyobjc-framework-ScreenSaver==8.3
pyobjc-framework-ScriptingBridge==8.3
pyobjc-framework-SearchKit==8.3
pyobjc-framework-Security==8.3
pyobjc-framework-SecurityFoundation==8.3
pyobjc-framework-SecurityInterface==8.3
pyobjc-framework-ServiceManagement==8.3
pyobjc-framework-Social==8.3
pyobjc-framework-SpriteKit==8.3
pyobjc-framework-StoreKit==8.3
pyobjc-framework-SyncServices==8.3
pyobjc-framework-SystemConfiguration==8.3
pyobjc-framework-VideoToolbox==8.3
pyobjc-framework-Vision==8.3
pyobjc-framework-WebKit==8.3
pyparsing==3.0.7
pytesseract==0.3.9
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
